+++ docs/create.py
@@ -53,8 +53,8 @@ README
 Introduction
 ============
 
-The main purpose of the PyPGx package is to provide a unified platform for
-pharmacogenomics (PGx) research.
+The main purpose of the PyPGx package, which is completely free and open
+source, is to provide a unified platform for pharmacogenomics (PGx) research.
 
 The package is written in Python, and supports both command line interface
 (CLI) and application programming interface (API) whose documentations are
@@ -185,6 +185,28 @@ you can access a development branch with the ``git checkout`` command. When
 you do this, please make sure your environment already has all the
 dependencies installed.
 
+Resource bundle
+===============
+
+Starting with the 0.12.0 version, reference haplotype panel files and
+structural variant classifier files in PyPGx are moved to the
+``pypgx-bundle`` `repository <https://github.com/sbslee/pypgx-bundle>`__
+(only those files are moved; other files such as ``allele-table.csv`` and
+``variant-table.csv`` are intact). Therefore, the user must clone the
+``pypgx-bundle`` repository with matching PyPGx version to their home
+directory in order for PyPGx to correctly access the moved files:
+
+.. code-block:: text
+
+   $ cd ~
+   $ git clone --branch 0.12.0 --depth 1 https://github.com/sbslee/pypgx-bundle
+
+This is undoubtedly annoying, but absolutely necessary for portability
+reasons because PyPGx has been growing exponentially in file size due to the
+increasing number of genes supported and their variation complexity, to the
+point where it now exceeds upload size limit for PyPI (100 Mb). After removal
+of those files, the size of PyPGx has reduced from >100 Mb to <1 Mb.
+
 Structural variation detection
 ==============================
 
@@ -205,7 +227,7 @@ class.OneVsRestClassifier.html>`__ for each gene for each GRCh build. Each
 classifier is trained using copy number profiles of real NGS samples as well
 as simulated ones.
 
-You can plot copy number profile and allele fraction profile with PyPGX to
+You can plot copy number profile and allele fraction profile with PyPGx to
 visually inspect SV calls. Below are CYP2D6 examples:
 
 .. list-table::
@@ -413,6 +435,19 @@ input and outputs a ``SampleTable[Phenotypes]`` file:
 
    $ pypgx call-phenotypes genotypes.zip phenotypes.zip
 
+Pipelines
+=========
+
+PyPGx provides two pipelines for performing PGx genotype analysis: NGS pipeline and chip pipeline.
+
+**NGS pipeline**
+
+.. image:: https://raw.githubusercontent.com/sbslee/pypgx-data/main/flowchart-ngs-pipeline.png
+
+**Chip pipeline**
+
+.. image:: https://raw.githubusercontent.com/sbslee/pypgx-data/main/flowchart-chip-pipeline.png
+
 Getting help
 ============
 
@@ -470,7 +505,7 @@ Above will print:
     SemanticType=CovFrame[DepthOfCoverage]
     Platform=WGS
 
-We can run the NGS pipeline for the *CYP2D6* gene:
+We can run the NGS pipeline for the CYP2D6 gene:
 
 .. code-block:: text
 
@@ -499,7 +534,7 @@ Above will create a number of archive files:
 API examples
 ============
 
-We can obtain allele function for the *CYP2D6* gene:
+We can obtain allele function for the CYP2D6 gene:
 
 .. code:: python3
 
@@ -513,7 +548,7 @@ We can obtain allele function for the *CYP2D6* gene:
     >>> pypgx.get_function('CYP2D6', '*140')
     'Unknown Function'
 
-We can predict phenotype for the *CYP2D6* gene based on two haplotype calls:
+We can predict phenotype for the CYP2D6 gene based on two haplotype calls:
 
 .. code:: python3