+++ pypgx/api/pipeline.py
@@ -31,8 +31,9 @@ def run_chip_pipeline(
     assembly : {'GRCh37', 'GRCh38'}, default: 'GRCh37'
         Reference genome assembly.
     panel : str, optional
-        VCF file corresponding to a reference haplotype panel (zipped or
-        unzipped). By default, the 1KGP panel is used.
+        VCF file corresponding to a reference haplotype panel (compressed or
+        uncompressed). By default, the 1KGP panel in the ``~/pypgx-bundle``
+        directory will be used.
     impute : bool, default: False
         If True, perform imputation of missing genotypes.
     force : bool, default : False
@@ -83,7 +84,8 @@ def run_ngs_pipeline(
     gene, output, variants=None, depth_of_coverage=None,
     control_statistics=None, platform='WGS', assembly='GRCh37', panel=None,
     force=False, samples=None, exclude=False, samples_without_sv=None,
-    do_not_plot_copy_number=False, do_not_plot_allele_fraction=False
+    do_not_plot_copy_number=False, do_not_plot_allele_fraction=False,
+    cnv_caller=None
 ):
     """
     Run PyPGx's genotyping pipeline for NGS data.
@@ -113,8 +115,9 @@ def run_ngs_pipeline(
     assembly : {'GRCh37', 'GRCh38'}, default: 'GRCh37'
         Reference genome assembly.
     panel : str, optional
-        VCF file corresponding to a reference haplotype panel (zipped or
-        unzipped). By default, the 1KGP panel is used.
+        VCF file corresponding to a reference haplotype panel (compressed or
+        uncompressed). By default, the 1KGP panel in the ``~/pypgx-bundle``
+        directory will be used.
     force : bool, default : False
         Overwrite output directory if it already exists.
     samples : str or list, optional
@@ -129,6 +132,10 @@ def run_ngs_pipeline(
         Do not plot copy number profile.
     do_not_plot_allele_fraction : bool, default: False
         Do not plot allele fraction profile.
+    cnv_caller : str or pypgx.Archive, optional
+        Archive file or object with the semantic type Model[CNV]. By default,
+        a pre-trained CNV caller in the ``~/pypgx-bundle`` directory will be
+        used.
     """
     if not core.is_target_gene(gene):
         raise core.NotTargetGeneError(gene)
@@ -224,7 +231,7 @@ def run_ngs_pipeline(
         copy_number = utils.compute_copy_number(read_depth,
             control_statistics, samples_without_sv=samples_without_sv)
         copy_number.to_file(f'{output}/copy-number.zip')
-        cnv_calls = utils.predict_cnv(copy_number)
+        cnv_calls = utils.predict_cnv(copy_number, cnv_caller=cnv_caller)
         cnv_calls.to_file(f'{output}/cnv-calls.zip')
         if not do_not_plot_copy_number:
             os.mkdir(f'{output}/copy-number-profile')